requirements.txt
streamlit
pandas
plotly
numpy
psycopg2-binary
